Implement dumping survey configuration to JSON
