+++ autowisp/browser_interface/configuration/views.py
@@ -1,11 +1,8 @@
 """The views to display and edit pipeline configuration."""
 
 from collections import namedtuple
-from traceback import print_exc
-import json
 
-from sqlalchemy import select, func, inspect, delete
-from sqlalchemy.orm import ColumnProperty
+from sqlalchemy import select, func, delete
 from django.shortcuts import render
 from django.http import HttpResponseRedirect
 from django.urls import reverse
@@ -14,17 +11,20 @@ from autowisp.database.user_interface import (
     get_json_config,
     save_json_config,
     list_steps,
+    get_editable_attributes,
+    update_db_entry,
+    get_human_name,
 )
 from autowisp.database.interface import start_db_session, set_sqlite_database
 
 # False positive
 # pylint: disable=no-name-in-module
 from autowisp.database.data_model import (
+    provenance,
     Configuration,
     ImageProcessingProgress,
     ObservingSession,
 )
-from autowisp.database.data_model import provenance
 
 # pylint: enable=no-name-in-module
 
@@ -67,50 +67,6 @@ def save_config(request, version):
     set_sqlite_database(request.session["project_db_path"])
     save_json_config(request.body, version)
     return HttpResponseRedirect(reverse("configuration:config_tree"))
-
-
-def get_human_name(column_name):
-    """Return human friendly name for the given column."""
-
-    if column_name == "serial_number":
-        return "serial no"
-    if column_name == "f_ratio":
-        return "focal ratio"
-    if column_name.endswith("_type_id"):
-        return "type"
-    return column_name.replace("_", " ")
-
-
-def get_editable_attributes(db_class):
-    """List the user-editable attributes for the given component DB class."""
-
-    def sort_key(colname):
-        """Define the order in which attributes should be displayed."""
-
-        if colname in ["name", "serial_number"]:
-            return 0
-        if colname == "type":
-            return 1
-        if colname == "notes":
-            return 3
-        return 2
-
-    columns = [
-        str(a).split(".", 1)[1]
-        for a in inspect(db_class).attrs
-        if isinstance(a, ColumnProperty)
-    ]
-    result = [
-        "type" if col_name.endswith("_type_id") else col_name
-        for col_name in columns
-        if col_name not in ["id", "timestamp"]
-    ]
-    if "type" in result:
-        result.remove("type")
-        result.append("type")
-    if db_class == provenance.CameraType:  # pylint: disable=no-member
-        result.append("channels")
-    return sorted(result, key=sort_key)
 
 
 def format_channel_attr(camera_type):
@@ -408,188 +364,6 @@ def delete_from_survey(
     return HttpResponseRedirect(reverse("configuration:survey"))
 
 
-def add_camera_type_channel(camera_type_id, properties, db_session):
-    """
-    Add channels to the given camera type and return partial channel entries.
-
-    Args:
-        camera_type_id(int):    The ID of the camera type to which to add
-            channels.
-
-        properties(dict-like):    The information being changed for the survey.
-            For each channel to add there should be exactly two keywords:
-            ``"channel-{channel_id}-name"`` and
-            ``"channel-{channel_id}-slice"``. Where ``{channel_id}`` should be
-            either an int specifying the identifier of the channel in the
-            database or ``"new"`` specifying a new channel to add.
-            ``{channel_id}``entries should be unique (for example only one new
-            channel can be added). Channel slices have the format:
-            ``"{x_offset}:{x_step};{y_offset}:{y_step}"``. Anything not related
-            to channels is ignored.
-
-        db_session:    The database session to use for updating.
-
-    Returns:
-        int or None, str or None:
-            The channel ID and property (one of ``"name"`` or ``"slice"``)
-            which is not fully specified or is mal-formatted. If more than one,
-            the one wit the lowest ID is returned. If the new channel is
-            unspecified, the channel returned is ``None``. If everything is
-            fully specified ``None, None`` is returned.
-    """
-
-    def get_channel_info():
-        """From the inputs extract the information to add to the database."""
-
-        result = {}
-        for key in properties:
-            if key.startswith("channel-"):
-                channel_id, channel_property = key.rsplit("-")[1:]
-                assert channel_property in [
-                    "name",
-                    "slice",
-                ], f"Unrecognized channel property {key}"
-                if channel_id != "new":
-                    channel_id = int(channel_id)
-                if channel_id not in result:
-                    result[channel_id] = {}
-                if channel_property == "name":
-                    assert "name" not in result[channel_id], (
-                        "Duplicate name entry encountered for channel ID "
-                        f"{channel_id}"
-                    )
-                    result[channel_id]["name"] = properties[key]
-                else:
-                    try:
-                        values = sum(
-                            (
-                                dir_slice.split(":")
-                                for dir_slice in properties[key].split(";")
-                            ),
-                            [],
-                        )
-                        values = [int(v) for v in values]
-                        for attr, val in zip(
-                            ["x_offset", "x_step", "y_offset", "y_step"], values
-                        ):
-                            assert attr not in result[channel_id], (
-                                "Duplicate slice entry encountered for channel "
-                                f"ID {channel_id}"
-                            )
-                            result[channel_id][attr] = val
-                    except ValueError:
-                        print_exc()
-        return result
-
-    def remove_unspecified(channel_info):
-        """Leave only fully specified channels in update info, return result."""
-
-        edit_id = None
-        edit_property = None
-        to_delete = set()
-        required_attributes = get_editable_attributes(
-            provenance.CameraChannel  # pylint: disable=no-member
-        )
-        required_attributes.remove("type")
-
-        for channel_id, channel_attrs in channel_info.items():
-            for attr in required_attributes:
-                if attr not in channel_attrs:
-                    print(
-                        f"Attribute {attr} mising. "
-                        f"Deleting channel {channel_id}."
-                    )
-                    if edit_id is None or edit_id > channel_id:
-                        edit_id = channel_id
-                        edit_property = "name" if attr == "name" else "slice"
-                    to_delete.add(channel_id)
-        for channel_id in to_delete:
-            del channel_info[channel_id]
-        return edit_id, edit_property
-
-    channel_info = get_channel_info()
-    print(80 * "*")
-    print(f"Channel info: {channel_info!r}")
-    result = remove_unspecified(channel_info)
-    print(f"Cleaned channel info: {channel_info!r}")
-    print(f"Result: {result!r}")
-    if channel_info:
-        assert (
-            camera_type_id >= 0
-        ), "Attempting to set channels of non-existant camera type"
-
-    for channel_id, channel_properties in channel_info.items():
-        print(f"Editing channel {channel_id} per: {channel_properties!r}")
-        if channel_id == "new":
-            db_channel = provenance.CameraChannel(  # pylint: disable=no-member
-                camera_type_id=camera_type_id, **channel_properties
-            )
-        else:
-            db_channel = db_session.scalar(
-                select(
-                    provenance.CameraChannel  # pylint: disable=no-member
-                ).filter_by(id=channel_id, camera_type_id=camera_type_id)
-            )
-            for attr, value in channel_properties.items():
-                setattr(db_channel, attr, value)
-
-        if channel_id == "new":
-            db_session.add(db_channel)
-    print(80 * "*")
-    return result
-
-
-def update_db_entry(
-    db_session, properties, db_class, entry_id, component_type=None
-):
-    """
-    Add/update a survey component or type, return its ID and what to autofocus.
-    """
-
-    print(80 * "*")
-    print(repr(properties))
-    print(80 * "*")
-
-    incomplete = None
-    entry_id = int(entry_id)
-    if entry_id < 0:
-        db_item = db_class()
-    else:
-        db_item = db_session.scalar(
-            select(db_class).where(db_class.id == entry_id)
-        )
-
-    attribute_names = get_editable_attributes(db_class)
-    for attr in attribute_names:
-        if attr == "channels":
-            assert (
-                db_class == provenance.CameraType  # pylint: disable=no-member
-            ), (
-                f"Attempting to set channels for {db_class} (not a camera "
-                "type)!"
-            )
-            channel_incomplete = add_camera_type_channel(
-                entry_id, properties, db_session
-            )
-            if (
-                channel_incomplete[0] is not None
-                or channel_incomplete[1] is not None
-            ):
-                incomplete = {"channel": channel_incomplete}
-        elif attr != "type":
-            setattr(db_item, attr, properties[get_human_name(attr)])
-
-    if "type" in attribute_names:
-        type_id = int(properties.get("type-id"))
-        assert type_id >= 0
-        setattr(db_item, component_type + "_type_id", type_id)
-
-    if entry_id < 0:
-        db_session.add(db_item)
-    db_session.commit()
-    return db_item.id, incomplete
-
-
 def update_survey_component_type(request, component_type, type_id):
     """Add or update a survey component type."""
 
@@ -632,64 +406,6 @@ def update_survey_component(request, component_type, component_id):
             component_type,
         )
     return HttpResponseRedirect(reverse("configuration:survey"))
-
-
-def import_json_to_survey(json_file):
-    """Add to the survey configuration from given JSON encoding string."""
-
-    config = json.load(json_file)
-    assert isinstance(
-        config, dict
-    ), "Malformatted JSON file encountered during import"
-
-    with start_db_session() as db_session:
-        for key, value in config.items():
-            key = key.title()
-            assert key.endswith(
-                "s"
-            ), f"Survey class {key} does not end with 's'."
-            if key in ["Observers", "Observatories"]:
-                db_class = (
-                    provenance.Observer  # pylint: disable=no-member
-                    if key == "Observers"
-                    else provenance.Observatory  # pylint: disable=no-member
-                )
-                incomplete = update_db_entry(db_session, value, db_class, -1)[1]
-            else:
-                component_type = key[:-1]
-                db_class = getattr(provenance, component_type + "Type")
-                type_id, incomplete = update_db_entry(
-                    db_session, value, db_class, -1
-                )
-                if incomplete:
-                    break
-
-                db_class = getattr(provenance, component_type)
-                for component in value["devices"]:
-                    component["type-id"] = type_id
-                    update_db_entry(
-                        db_session,
-                        component,
-                        db_class,
-                        -1,
-                        component_type.lower(),
-                    )
-                if component_type == "Camera":
-                    for channel_name, channel_config in value["channels"]:
-                        channel_config["name"] = channel_name
-                        incomplete = update_db_entry(
-                            db_session,
-                            channel_config,
-                            provenance.CameraChannel,  # pylint: disable=no-member
-                            -1,
-                            "camera",
-                        )[1]
-                        if incomplete:
-                            break
-            assert incomplete is None, (
-                "Mal-formatted or not fully specified configuration for "
-                f"{key}: {value!r}"
-            )
 
 
 def change_access(  # pylint: disable=too-many-positional-arguments too-many-arguments line-too-long
